Agent simple/requirements.txt
# ─────────────────────────────────────────────────────────────────
#  requirements.txt  —  LangChain Predictive Maintenance Agent
#  Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────

# ── Core LangChain ──────────────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# ── UI & Visualisation ──────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0

# ── HTTP & Environment ──────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
